rework to separate out python module
Remove the magic from the magic _db* modules.
Combine as one 'devsup._dbapi'.

libpyDevSup exists only to bootstrap python
interpreter in IOCs.
