+++ pyIocApp/pyDevSup.dbd
@@ -1,0 +1,1 @@
+registrar(pySetupReg)
